+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Fix error with automatic `highclip` or `lowclip` and scalar colors [#4048](https://github.com/MakieOrg/Makie.jl/pull/4048).
 - Correct a bug in the `project` function when projecting using a `Scene`. [#3909](https://github.com/MakieOrg/Makie.jl/pull/3909).
 - Correct a method ambiguity in `insert!` which was causing `PlotList` to fail on CairoMakie. [#4038](https://github.com/MakieOrg/Makie.jl/pull/4038)
+- Prepare better support for Wayland [#4062](https://github.com/MakieOrg/Makie.jl/pull/4062)
 
 ## [0.21.5] - 2024-07-07